+++ include/D3D12Helper/D3D12Gpu/D3D12Gpu.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <D3D12Helper/D3D12Gpu/D3D12Binding.hpp>
 #include <D3D12Helper/D3D12Gpu/D3D12CpuImage.hpp>
 #include <D3D12Helper/D3D12Gpu/D3D12TextureTransfer.hpp>
 #include <D3D12Helper/D3D12Framework/D3D12Resource.hpp>